@@ -1,6 +1,6 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-from typing import Optional
+from typing import List, Optional
 
 from .._models import BaseModel
 
@@ -10,4 +10,4 @@ __all__ = ["SwarmCheckAvailableResponse"]
 class SwarmCheckAvailableResponse(BaseModel):
     success: Optional[bool] = None
 
-    swarm_types: Optional[object] = None
+    swarm_types: Optional[List[str]] = None
