@@ -9,7 +9,7 @@ import pytest
 
 from tests.utils import assert_matches_type
 from swarms_client import SwarmsClient, AsyncSwarmsClient
-from swarms_client.types import AgentRunResponse
+from swarms_client.types import AgentRunResponse, AgentListResponse
 
 base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
 
@@ -18,6 +18,34 @@ class TestAgent:
     parametrize = pytest.mark.parametrize(
         "client", [False, True], indirect=True, ids=["loose", "strict"]
     )
+
+    @pytest.mark.skip(reason="Prism tests are disabled")
+    @parametrize
+    def test_method_list(self, client: SwarmsClient) -> None:
+        agent = client.agent.list()
+        assert_matches_type(AgentListResponse, agent, path=["response"])
+
+    @pytest.mark.skip(reason="Prism tests are disabled")
+    @parametrize
+    def test_raw_response_list(self, client: SwarmsClient) -> None:
+        response = client.agent.with_raw_response.list()
+
+        assert response.is_closed is True
+        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+        agent = response.parse()
+        assert_matches_type(AgentListResponse, agent, path=["response"])
+
+    @pytest.mark.skip(reason="Prism tests are disabled")
+    @parametrize
+    def test_streaming_response_list(self, client: SwarmsClient) -> None:
+        with client.agent.with_streaming_response.list() as response:
+            assert not response.is_closed
+            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+
+            agent = response.parse()
+            assert_matches_type(AgentListResponse, agent, path=["response"])
+
+        assert cast(Any, response.is_closed) is True
 
     @pytest.mark.skip(reason="Prism tests are disabled")
     @parametrize
@@ -109,6 +137,34 @@ class TestAsyncAgent:
         indirect=True,
         ids=["loose", "strict", "aiohttp"],
     )
+
+    @pytest.mark.skip(reason="Prism tests are disabled")
+    @parametrize
+    async def test_method_list(self, async_client: AsyncSwarmsClient) -> None:
+        agent = await async_client.agent.list()
+        assert_matches_type(AgentListResponse, agent, path=["response"])
+
+    @pytest.mark.skip(reason="Prism tests are disabled")
+    @parametrize
+    async def test_raw_response_list(self, async_client: AsyncSwarmsClient) -> None:
+        response = await async_client.agent.with_raw_response.list()
+
+        assert response.is_closed is True
+        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+        agent = await response.parse()
+        assert_matches_type(AgentListResponse, agent, path=["response"])
+
+    @pytest.mark.skip(reason="Prism tests are disabled")
+    @parametrize
+    async def test_streaming_response_list(self, async_client: AsyncSwarmsClient) -> None:
+        async with async_client.agent.with_streaming_response.list() as response:
+            assert not response.is_closed
+            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+
+            agent = await response.parse()
+            assert_matches_type(AgentListResponse, agent, path=["response"])
+
+        assert cast(Any, response.is_closed) is True
 
     @pytest.mark.skip(reason="Prism tests are disabled")
     @parametrize
