@@ -65,23 +65,21 @@ def _get_open_connections(client: SwarmsClient | AsyncSwarmsClient) -> int:
 
 
 class TestSwarmsClient:
-    client = SwarmsClient(
-        base_url=base_url, api_key=api_key, _strict_response_validation=True
-    )
+    
 
     @pytest.mark.respx(base_url=base_url)
-    def test_raw_response(self, respx_mock: MockRouter) -> None:
+    def test_raw_response(self, respx_mock: MockRouter, client: SwarmsClient) -> None:
         respx_mock.post("/foo").mock(
             return_value=httpx.Response(200, json={"foo": "bar"})
         )
 
-        response = self.client.post("/foo", cast_to=httpx.Response)
+        response = client.post("/foo", cast_to=httpx.Response)
         assert response.status_code == 200
         assert isinstance(response, httpx.Response)
         assert response.json() == {"foo": "bar"}
 
     @pytest.mark.respx(base_url=base_url)
-    def test_raw_response_for_binary(self, respx_mock: MockRouter) -> None:
+    def test_raw_response_for_binary(self, respx_mock: MockRouter, client: SwarmsClient) -> None:
         respx_mock.post("/foo").mock(
             return_value=httpx.Response(
                 200,
@@ -90,34 +88,34 @@ class TestSwarmsClient:
             )
         )
 
-        response = self.client.post("/foo", cast_to=httpx.Response)
+        response = client.post("/foo", cast_to=httpx.Response)
         assert response.status_code == 200
         assert isinstance(response, httpx.Response)
         assert response.json() == {"foo": "bar"}
 
-    def test_copy(self) -> None:
-        copied = self.client.copy()
-        assert id(copied) != id(self.client)
+    def test_copy(self, client: SwarmsClient) -> None:
+        copied = client.copy()
+        assert id(copied) != id(client)
 
-        copied = self.client.copy(api_key="another My API Key")
+        copied = client.copy(api_key="another My API Key")
         assert copied.api_key == "another My API Key"
-        assert self.client.api_key == "My API Key"
+        assert client.api_key == "My API Key"
 
-    def test_copy_default_options(self) -> None:
+    def test_copy_default_options(self, client: SwarmsClient) -> None:
         # options that have a default are overridden correctly
-        copied = self.client.copy(max_retries=7)
+        copied = client.copy(max_retries=7)
         assert copied.max_retries == 7
-        assert self.client.max_retries == 2
+        assert client.max_retries == 2
 
         copied2 = copied.copy(max_retries=6)
         assert copied2.max_retries == 6
         assert copied.max_retries == 7
 
         # timeout
-        assert isinstance(self.client.timeout, httpx.Timeout)
-        copied = self.client.copy(timeout=None)
+        assert isinstance(client.timeout, httpx.Timeout)
+        copied = client.copy(timeout=None)
         assert copied.timeout is None
-        assert isinstance(self.client.timeout, httpx.Timeout)
+        assert isinstance(client.timeout, httpx.Timeout)
 
     def test_copy_default_headers(self) -> None:
         client = SwarmsClient(
@@ -155,6 +153,7 @@ class TestSwarmsClient:
             match="`default_headers` and `set_default_headers` arguments are mutually exclusive",
         ):
             client.copy(set_default_headers={}, default_headers={"X-Foo": "Bar"})
+        client.close()
 
     def test_copy_default_query(self) -> None:
         client = SwarmsClient(
@@ -195,13 +194,15 @@ class TestSwarmsClient:
         ):
             client.copy(set_default_query={}, default_query={"foo": "Bar"})
 
-    def test_copy_signature(self) -> None:
+        client.close()
+
+    def test_copy_signature(self, client: SwarmsClient) -> None:
         # ensure the same parameters that can be passed to the client are defined in the `.copy()` method
         init_signature = inspect.signature(
             # mypy doesn't like that we access the `__init__` property.
-            self.client.__init__,  # type: ignore[misc]
+            client.__init__,  # type: ignore[misc]
         )
-        copy_signature = inspect.signature(self.client.copy)
+        copy_signature = inspect.signature(client.copy)
         exclude_params = {"transport", "proxies", "_strict_response_validation"}
 
         for name in init_signature.parameters.keys():
@@ -217,12 +218,12 @@ class TestSwarmsClient:
         sys.version_info >= (3, 10),
         reason="fails because of a memory leak that started from 3.12",
     )
-    def test_copy_build_request(self) -> None:
+    def test_copy_build_request(self, client: SwarmsClient) -> None:
         options = FinalRequestOptions(method="get", url="/foo")
 
         def build_request(options: FinalRequestOptions) -> None:
-            client = self.client.copy()
-            client._build_request(options)
+            client_copy = client.copy()
+            client_copy._build_request(options)
 
         # ensure that the machinery is warmed up before tracing starts.
         build_request(options)
@@ -281,16 +282,14 @@ class TestSwarmsClient:
                     print(frame)
             raise AssertionError()
 
-    def test_request_timeout(self) -> None:
-        request = self.client._build_request(
+    def test_request_timeout(self, client: SwarmsClient) -> None:
+        request = client._build_request(
             FinalRequestOptions(method="get", url="/foo")
         )
         timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
         assert timeout == DEFAULT_TIMEOUT
 
-        request = self.client._build_request(
-            FinalRequestOptions(method="get", url="/foo", timeout=httpx.Timeout(100.0))
-        )
+        request = client._build_request(FinalRequestOptions(method="get", url="/foo", timeout=httpx.Timeout(100.0)))
         timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
         assert timeout == httpx.Timeout(100.0)
 
@@ -305,6 +304,8 @@ class TestSwarmsClient:
         request = client._build_request(FinalRequestOptions(method="get", url="/foo"))
         timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
         assert timeout == httpx.Timeout(0)
+
+        client.close()
 
     def test_http_client_timeout_option(self) -> None:
         # custom timeout given to the httpx client should be used
@@ -322,6 +323,8 @@ class TestSwarmsClient:
             timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
             assert timeout == httpx.Timeout(None)
 
+            client.close()
+
         # no timeout given to the httpx client should not use the httpx default
         with httpx.Client() as http_client:
             client = SwarmsClient(
@@ -336,6 +339,8 @@ class TestSwarmsClient:
             )
             timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
             assert timeout == DEFAULT_TIMEOUT
+
+            client.close()
 
         # explicitly passing the default timeout currently results in it being ignored
         with httpx.Client(timeout=HTTPX_DEFAULT_TIMEOUT) as http_client:
@@ -352,6 +357,8 @@ class TestSwarmsClient:
             timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
             assert timeout == DEFAULT_TIMEOUT  # our default
 
+            client.close()
+
     async def test_invalid_http_client(self) -> None:
         with pytest.raises(TypeError, match="Invalid `http_client` arg"):
             async with httpx.AsyncClient() as http_client:
@@ -363,17 +370,17 @@ class TestSwarmsClient:
                 )
 
     def test_default_headers_option(self) -> None:
-        client = SwarmsClient(
+        test_client = SwarmsClient(
             base_url=base_url,
             api_key=api_key,
             _strict_response_validation=True,
             default_headers={"X-Foo": "bar"},
         )
-        request = client._build_request(FinalRequestOptions(method="get", url="/foo"))
+        request = test_client._build_request(FinalRequestOptions(method="get", url="/foo"))
         assert request.headers.get("x-foo") == "bar"
         assert request.headers.get("x-stainless-lang") == "python"
 
-        client2 = SwarmsClient(
+        test_client2 = SwarmsClient(
             base_url=base_url,
             api_key=api_key,
             _strict_response_validation=True,
@@ -382,9 +389,12 @@ class TestSwarmsClient:
                 "X-Stainless-Lang": "my-overriding-header",
             },
         )
-        request = client2._build_request(FinalRequestOptions(method="get", url="/foo"))
+        request = test_client2._build_request(FinalRequestOptions(method="get", url="/foo"))
         assert request.headers.get("x-foo") == "stainless"
         assert request.headers.get("x-stainless-lang") == "my-overriding-header"
+
+        test_client.close()
+        test_client2.close()
 
     def test_validate_headers(self) -> None:
         client = SwarmsClient(
@@ -430,8 +440,10 @@ class TestSwarmsClient:
         url = httpx.URL(request.url)
         assert dict(url.params) == {"foo": "baz", "query_param": "overridden"}
 
-    def test_request_extra_json(self) -> None:
-        request = self.client._build_request(
+        client.close()
+
+    def test_request_extra_json(self, client: SwarmsClient) -> None:
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -442,7 +454,7 @@ class TestSwarmsClient:
         data = json.loads(request.content.decode("utf-8"))
         assert data == {"foo": "bar", "baz": False}
 
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -453,7 +465,7 @@ class TestSwarmsClient:
         assert data == {"baz": False}
 
         # `extra_json` takes priority over `json_data` when keys clash
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -464,8 +476,8 @@ class TestSwarmsClient:
         data = json.loads(request.content.decode("utf-8"))
         assert data == {"foo": "bar", "baz": None}
 
-    def test_request_extra_headers(self) -> None:
-        request = self.client._build_request(
+    def test_request_extra_headers(self, client: SwarmsClient) -> None:
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -475,7 +487,7 @@ class TestSwarmsClient:
         assert request.headers.get("X-Foo") == "Foo"
 
         # `extra_headers` takes priority over `default_headers` when keys clash
-        request = self.client.with_options(
+        request = client.with_options(
             default_headers={"X-Bar": "true"}
         )._build_request(
             FinalRequestOptions(
@@ -488,8 +500,8 @@ class TestSwarmsClient:
         )
         assert request.headers.get("X-Bar") == "false"
 
-    def test_request_extra_query(self) -> None:
-        request = self.client._build_request(
+    def test_request_extra_query(self, client: SwarmsClient) -> None:
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -502,7 +514,7 @@ class TestSwarmsClient:
         assert params == {"my_query_param": "Foo"}
 
         # if both `query` and `extra_query` are given, they are merged
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -516,7 +528,7 @@ class TestSwarmsClient:
         assert params == {"bar": "1", "foo": "2"}
 
         # `extra_query` takes priority over `query` when keys clash
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -561,7 +573,7 @@ class TestSwarmsClient:
         ]
 
     @pytest.mark.respx(base_url=base_url)
-    def test_basic_union_response(self, respx_mock: MockRouter) -> None:
+    def test_basic_union_response(self, respx_mock: MockRouter, client: SwarmsClient) -> None:
         class Model1(BaseModel):
             name: str
 
@@ -572,12 +584,12 @@ class TestSwarmsClient:
             return_value=httpx.Response(200, json={"foo": "bar"})
         )
 
-        response = self.client.get("/foo", cast_to=cast(Any, Union[Model1, Model2]))
+        response = client.get("/foo", cast_to=cast(Any, Union[Model1, Model2]))
         assert isinstance(response, Model2)
         assert response.foo == "bar"
 
     @pytest.mark.respx(base_url=base_url)
-    def test_union_response_different_types(self, respx_mock: MockRouter) -> None:
+    def test_union_response_different_types(self, respx_mock: MockRouter, client: SwarmsClient) -> None:
         """Union of objects with the same field name using a different type"""
 
         class Model1(BaseModel):
@@ -590,19 +602,19 @@ class TestSwarmsClient:
             return_value=httpx.Response(200, json={"foo": "bar"})
         )
 
-        response = self.client.get("/foo", cast_to=cast(Any, Union[Model1, Model2]))
+        response = client.get("/foo", cast_to=cast(Any, Union[Model1, Model2]))
         assert isinstance(response, Model2)
         assert response.foo == "bar"
 
         respx_mock.get("/foo").mock(return_value=httpx.Response(200, json={"foo": 1}))
 
-        response = self.client.get("/foo", cast_to=cast(Any, Union[Model1, Model2]))
+        response = client.get("/foo", cast_to=cast(Any, Union[Model1, Model2]))
         assert isinstance(response, Model1)
         assert response.foo == 1
 
     @pytest.mark.respx(base_url=base_url)
     def test_non_application_json_content_type_for_json_data(
-        self, respx_mock: MockRouter
+        self, respx_mock: MockRouter, client: SwarmsClient
     ) -> None:
         """
         Response that sets Content-Type to something other than application/json but returns json data
@@ -619,7 +631,7 @@ class TestSwarmsClient:
             )
         )
 
-        response = self.client.get("/foo", cast_to=Model)
+        response = client.get("/foo", cast_to=Model)
         assert isinstance(response, Model)
         assert response.foo == 2
 
@@ -635,10 +647,24 @@ class TestSwarmsClient:
 
         assert client.base_url == "https://example.com/from_setter/"
 
+        client.close()
+
     def test_base_url_env(self) -> None:
         with update_env(SWARMS_CLIENT_BASE_URL="http://localhost:5000/from/env"):
             client = SwarmsClient(api_key=api_key, _strict_response_validation=True)
             assert client.base_url == "http://localhost:5000/from/env/"
+
+        # explicit environment arg requires explicitness
+        with update_env(SWARMS_CLIENT_BASE_URL="http://localhost:5000/from/env"):
+            with pytest.raises(ValueError, match=r"you must pass base_url=None"):
+                SwarmsClient(api_key=api_key, _strict_response_validation=True, environment="production")
+
+            client = SwarmsClient(
+                base_url=None, api_key=api_key, _strict_response_validation=True, environment="production"
+            )
+            assert str(client.base_url).startswith("https://api.swarms.world")
+
+            client.close()
 
     @pytest.mark.parametrize(
         "client",
@@ -666,6 +692,7 @@ class TestSwarmsClient:
             ),
         )
         assert request.url == "http://localhost:5000/custom/path/foo"
+        client.close()
 
     @pytest.mark.parametrize(
         "client",
@@ -693,6 +720,7 @@ class TestSwarmsClient:
             ),
         )
         assert request.url == "http://localhost:5000/custom/path/foo"
+        client.close()
 
     @pytest.mark.parametrize(
         "client",
@@ -720,32 +748,33 @@ class TestSwarmsClient:
             ),
         )
         assert request.url == "https://myapi.com/foo"
+        client.close()
 
     def test_copied_client_does_not_close_http(self) -> None:
-        client = SwarmsClient(
+        test_client = SwarmsClient(
             base_url=base_url, api_key=api_key, _strict_response_validation=True
         )
-        assert not client.is_closed()
+        assert not test_client.is_closed()
 
-        copied = client.copy()
-        assert copied is not client
+        copied = test_client.copy()
+        assert copied is not test_client
 
         del copied
 
-        assert not client.is_closed()
+        assert not test_client.is_closed()
 
     def test_client_context_manager(self) -> None:
-        client = SwarmsClient(
+        test_client = SwarmsClient(
             base_url=base_url, api_key=api_key, _strict_response_validation=True
         )
-        with client as c2:
-            assert c2 is client
+        with test_client as c2:
+            assert c2 is test_client
             assert not c2.is_closed()
-            assert not client.is_closed()
-        assert client.is_closed()
+            assert not test_client.is_closed()
+        assert test_client.is_closed()
 
     @pytest.mark.respx(base_url=base_url)
-    def test_client_response_validation_error(self, respx_mock: MockRouter) -> None:
+    def test_client_response_validation_error(self, respx_mock: MockRouter, client: SwarmsClient) -> None:
         class Model(BaseModel):
             foo: str
 
@@ -754,7 +783,7 @@ class TestSwarmsClient:
         )
 
         with pytest.raises(APIResponseValidationError) as exc:
-            self.client.get("/foo", cast_to=Model)
+            client.get("/foo", cast_to=Model)
 
         assert isinstance(exc.value.__cause__, ValidationError)
 
@@ -783,12 +812,15 @@ class TestSwarmsClient:
         with pytest.raises(APIResponseValidationError):
             strict_client.get("/foo", cast_to=Model)
 
-        client = SwarmsClient(
+        non_strict_client = SwarmsClient(
             base_url=base_url, api_key=api_key, _strict_response_validation=False
         )
 
-        response = client.get("/foo", cast_to=Model)
+        response = non_strict_client.get("/foo", cast_to=Model)
         assert isinstance(response, str)  # type: ignore[unreachable]
+
+        strict_client.close()
+        non_strict_client.close()
 
     @pytest.mark.parametrize(
         "remaining_retries,retry_after,timeout",
@@ -813,11 +845,9 @@ class TestSwarmsClient:
     )
     @mock.patch("time.time", mock.MagicMock(return_value=1696004797))
     def test_parse_retry_after_header(
-        self, remaining_retries: int, retry_after: str, timeout: float
+        self, remaining_retries: int, retry_after: str, timeout: float, client: SwarmsClient
     ) -> None:
-        client = SwarmsClient(
-            base_url=base_url, api_key=api_key, _strict_response_validation=True
-        )
+        
 
         headers = httpx.Headers({"retry-after": retry_after})
         options = FinalRequestOptions(method="get", url="/foo", max_retries=3)
@@ -843,7 +873,7 @@ class TestSwarmsClient:
         with pytest.raises(APITimeoutError):
             client.with_streaming_response.get_root().__enter__()
 
-        assert _get_open_connections(self.client) == 0
+        assert _get_open_connections(client) == 0
 
     @mock.patch(
         "swarms_client._base_client.BaseClient._calculate_retry_timeout",
@@ -857,7 +887,7 @@ class TestSwarmsClient:
 
         with pytest.raises(APIStatusError):
             client.with_streaming_response.get_root().__enter__()
-        assert _get_open_connections(self.client) == 0
+        assert _get_open_connections(client) == 0
 
     @pytest.mark.parametrize("failures_before_success", [0, 2, 4])
     @mock.patch(
@@ -977,7 +1007,7 @@ class TestSwarmsClient:
         )
 
     @pytest.mark.respx(base_url=base_url)
-    def test_follow_redirects(self, respx_mock: MockRouter) -> None:
+    def test_follow_redirects(self, respx_mock: MockRouter, client: SwarmsClient) -> None:
         # Test that the default follow_redirects=True allows following redirects
         respx_mock.post("/redirect").mock(
             return_value=httpx.Response(
@@ -988,14 +1018,14 @@ class TestSwarmsClient:
             return_value=httpx.Response(200, json={"status": "ok"})
         )
 
-        response = self.client.post(
+        response = client.post(
             "/redirect", body={"key": "value"}, cast_to=httpx.Response
         )
         assert response.status_code == 200
         assert response.json() == {"status": "ok"}
 
     @pytest.mark.respx(base_url=base_url)
-    def test_follow_redirects_disabled(self, respx_mock: MockRouter) -> None:
+    def test_follow_redirects_disabled(self, respx_mock: MockRouter, client: SwarmsClient) -> None:
         # Test that follow_redirects=False prevents following redirects
         respx_mock.post("/redirect").mock(
             return_value=httpx.Response(
@@ -1004,8 +1034,7 @@ class TestSwarmsClient:
         )
 
         with pytest.raises(APIStatusError) as exc_info:
-            self.client.post(
-                "/redirect",
+            client.post("/redirect",
                 body={"key": "value"},
                 options={"follow_redirects": False},
                 cast_to=httpx.Response,
@@ -1016,25 +1045,21 @@ class TestSwarmsClient:
 
 
 class TestAsyncSwarmsClient:
-    client = AsyncSwarmsClient(
-        base_url=base_url, api_key=api_key, _strict_response_validation=True
-    )
+    
 
     @pytest.mark.respx(base_url=base_url)
-    @pytest.mark.asyncio
-    async def test_raw_response(self, respx_mock: MockRouter) -> None:
+    async def test_raw_response(self, respx_mock: MockRouter, async_client: AsyncSwarmsClient) -> None:
         respx_mock.post("/foo").mock(
             return_value=httpx.Response(200, json={"foo": "bar"})
         )
 
-        response = await self.client.post("/foo", cast_to=httpx.Response)
+        response = await async_client.post("/foo", cast_to=httpx.Response)
         assert response.status_code == 200
         assert isinstance(response, httpx.Response)
         assert response.json() == {"foo": "bar"}
 
     @pytest.mark.respx(base_url=base_url)
-    @pytest.mark.asyncio
-    async def test_raw_response_for_binary(self, respx_mock: MockRouter) -> None:
+    async def test_raw_response_for_binary(self, respx_mock: MockRouter, async_client: AsyncSwarmsClient) -> None:
         respx_mock.post("/foo").mock(
             return_value=httpx.Response(
                 200,
@@ -1043,36 +1068,36 @@ class TestAsyncSwarmsClient:
             )
         )
 
-        response = await self.client.post("/foo", cast_to=httpx.Response)
+        response = await async_client.post("/foo", cast_to=httpx.Response)
         assert response.status_code == 200
         assert isinstance(response, httpx.Response)
         assert response.json() == {"foo": "bar"}
 
-    def test_copy(self) -> None:
-        copied = self.client.copy()
-        assert id(copied) != id(self.client)
+    def test_copy(self, async_client: AsyncSwarmsClient) -> None:
+        copied = async_client.copy()
+        assert id(copied) != id(async_client)
 
-        copied = self.client.copy(api_key="another My API Key")
+        copied = async_client.copy(api_key="another My API Key")
         assert copied.api_key == "another My API Key"
-        assert self.client.api_key == "My API Key"
+        assert async_client.api_key == "My API Key"
 
-    def test_copy_default_options(self) -> None:
+    def test_copy_default_options(self, async_client: AsyncSwarmsClient) -> None:
         # options that have a default are overridden correctly
-        copied = self.client.copy(max_retries=7)
+        copied = async_client.copy(max_retries=7)
         assert copied.max_retries == 7
-        assert self.client.max_retries == 2
+        assert async_client.max_retries == 2
 
         copied2 = copied.copy(max_retries=6)
         assert copied2.max_retries == 6
         assert copied.max_retries == 7
 
         # timeout
-        assert isinstance(self.client.timeout, httpx.Timeout)
-        copied = self.client.copy(timeout=None)
+        assert isinstance(async_client.timeout, httpx.Timeout)
+        copied = async_client.copy(timeout=None)
         assert copied.timeout is None
-        assert isinstance(self.client.timeout, httpx.Timeout)
+        assert isinstance(async_client.timeout, httpx.Timeout)
 
-    def test_copy_default_headers(self) -> None:
+    async def test_copy_default_headers(self) -> None:
         client = AsyncSwarmsClient(
             base_url=base_url,
             api_key=api_key,
@@ -1108,8 +1133,9 @@ class TestAsyncSwarmsClient:
             match="`default_headers` and `set_default_headers` arguments are mutually exclusive",
         ):
             client.copy(set_default_headers={}, default_headers={"X-Foo": "Bar"})
+        await client.close()
 
-    def test_copy_default_query(self) -> None:
+    async def test_copy_default_query(self) -> None:
         client = AsyncSwarmsClient(
             base_url=base_url,
             api_key=api_key,
@@ -1148,13 +1174,15 @@ class TestAsyncSwarmsClient:
         ):
             client.copy(set_default_query={}, default_query={"foo": "Bar"})
 
-    def test_copy_signature(self) -> None:
+        await client.close()
+
+    def test_copy_signature(self, async_client: AsyncSwarmsClient) -> None:
         # ensure the same parameters that can be passed to the client are defined in the `.copy()` method
         init_signature = inspect.signature(
             # mypy doesn't like that we access the `__init__` property.
-            self.client.__init__,  # type: ignore[misc]
+            async_client.__init__,  # type: ignore[misc]
         )
-        copy_signature = inspect.signature(self.client.copy)
+        copy_signature = inspect.signature(async_client.copy)
         exclude_params = {"transport", "proxies", "_strict_response_validation"}
 
         for name in init_signature.parameters.keys():
@@ -1170,12 +1198,12 @@ class TestAsyncSwarmsClient:
         sys.version_info >= (3, 10),
         reason="fails because of a memory leak that started from 3.12",
     )
-    def test_copy_build_request(self) -> None:
+    def test_copy_build_request(self, async_client: AsyncSwarmsClient) -> None:
         options = FinalRequestOptions(method="get", url="/foo")
 
         def build_request(options: FinalRequestOptions) -> None:
-            client = self.client.copy()
-            client._build_request(options)
+            client_copy = async_client.copy()
+            client_copy._build_request(options)
 
         # ensure that the machinery is warmed up before tracing starts.
         build_request(options)
@@ -1234,14 +1262,14 @@ class TestAsyncSwarmsClient:
                     print(frame)
             raise AssertionError()
 
-    async def test_request_timeout(self) -> None:
-        request = self.client._build_request(
+    async def test_request_timeout(self, async_client: AsyncSwarmsClient) -> None:
+        request = async_client._build_request(
             FinalRequestOptions(method="get", url="/foo")
         )
         timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
         assert timeout == DEFAULT_TIMEOUT
 
-        request = self.client._build_request(
+        request = async_client._build_request(
             FinalRequestOptions(method="get", url="/foo", timeout=httpx.Timeout(100.0))
         )
         timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
@@ -1259,6 +1287,8 @@ class TestAsyncSwarmsClient:
         timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
         assert timeout == httpx.Timeout(0)
 
+        await client.close()
+
     async def test_http_client_timeout_option(self) -> None:
         # custom timeout given to the httpx client should be used
         async with httpx.AsyncClient(timeout=None) as http_client:
@@ -1275,6 +1305,8 @@ class TestAsyncSwarmsClient:
             timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
             assert timeout == httpx.Timeout(None)
 
+            await client.close()
+
         # no timeout given to the httpx client should not use the httpx default
         async with httpx.AsyncClient() as http_client:
             client = AsyncSwarmsClient(
@@ -1289,6 +1321,8 @@ class TestAsyncSwarmsClient:
             )
             timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
             assert timeout == DEFAULT_TIMEOUT
+
+            await client.close()
 
         # explicitly passing the default timeout currently results in it being ignored
         async with httpx.AsyncClient(timeout=HTTPX_DEFAULT_TIMEOUT) as http_client:
@@ -1305,6 +1339,8 @@ class TestAsyncSwarmsClient:
             timeout = httpx.Timeout(**request.extensions["timeout"])  # type: ignore
             assert timeout == DEFAULT_TIMEOUT  # our default
 
+            await client.close()
+
     def test_invalid_http_client(self) -> None:
         with pytest.raises(TypeError, match="Invalid `http_client` arg"):
             with httpx.Client() as http_client:
@@ -1315,18 +1351,18 @@ class TestAsyncSwarmsClient:
                     http_client=cast(Any, http_client),
                 )
 
-    def test_default_headers_option(self) -> None:
-        client = AsyncSwarmsClient(
+    async def test_default_headers_option(self) -> None:
+        test_client = AsyncSwarmsClient(
             base_url=base_url,
             api_key=api_key,
             _strict_response_validation=True,
             default_headers={"X-Foo": "bar"},
         )
-        request = client._build_request(FinalRequestOptions(method="get", url="/foo"))
+        request = test_client._build_request(FinalRequestOptions(method="get", url="/foo"))
         assert request.headers.get("x-foo") == "bar"
         assert request.headers.get("x-stainless-lang") == "python"
 
-        client2 = AsyncSwarmsClient(
+        test_client2 = AsyncSwarmsClient(
             base_url=base_url,
             api_key=api_key,
             _strict_response_validation=True,
@@ -1335,9 +1371,12 @@ class TestAsyncSwarmsClient:
                 "X-Stainless-Lang": "my-overriding-header",
             },
         )
-        request = client2._build_request(FinalRequestOptions(method="get", url="/foo"))
+        request = test_client2._build_request(FinalRequestOptions(method="get", url="/foo"))
         assert request.headers.get("x-foo") == "stainless"
         assert request.headers.get("x-stainless-lang") == "my-overriding-header"
+
+        await test_client.close()
+        await test_client2.close()
 
     def test_validate_headers(self) -> None:
         client = AsyncSwarmsClient(
@@ -1362,7 +1401,7 @@ class TestAsyncSwarmsClient:
         )
         assert request2.headers.get("x-api-key") is None
 
-    def test_default_query_option(self) -> None:
+    async def test_default_query_option(self) -> None:
         client = AsyncSwarmsClient(
             base_url=base_url,
             api_key=api_key,
@@ -1383,8 +1422,10 @@ class TestAsyncSwarmsClient:
         url = httpx.URL(request.url)
         assert dict(url.params) == {"foo": "baz", "query_param": "overridden"}
 
-    def test_request_extra_json(self) -> None:
-        request = self.client._build_request(
+        await client.close()
+
+    def test_request_extra_json(self, client: SwarmsClient) -> None:
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -1395,7 +1436,7 @@ class TestAsyncSwarmsClient:
         data = json.loads(request.content.decode("utf-8"))
         assert data == {"foo": "bar", "baz": False}
 
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -1406,7 +1447,7 @@ class TestAsyncSwarmsClient:
         assert data == {"baz": False}
 
         # `extra_json` takes priority over `json_data` when keys clash
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -1417,8 +1458,8 @@ class TestAsyncSwarmsClient:
         data = json.loads(request.content.decode("utf-8"))
         assert data == {"foo": "bar", "baz": None}
 
-    def test_request_extra_headers(self) -> None:
-        request = self.client._build_request(
+    def test_request_extra_headers(self, client: SwarmsClient) -> None:
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -1428,7 +1469,7 @@ class TestAsyncSwarmsClient:
         assert request.headers.get("X-Foo") == "Foo"
 
         # `extra_headers` takes priority over `default_headers` when keys clash
-        request = self.client.with_options(
+        request = client.with_options(
             default_headers={"X-Bar": "true"}
         )._build_request(
             FinalRequestOptions(
@@ -1441,8 +1482,8 @@ class TestAsyncSwarmsClient:
         )
         assert request.headers.get("X-Bar") == "false"
 
-    def test_request_extra_query(self) -> None:
-        request = self.client._build_request(
+    def test_request_extra_query(self, client: SwarmsClient) -> None:
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -1455,7 +1496,7 @@ class TestAsyncSwarmsClient:
         assert params == {"my_query_param": "Foo"}
 
         # if both `query` and `extra_query` are given, they are merged
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -1469,7 +1510,7 @@ class TestAsyncSwarmsClient:
         assert params == {"bar": "1", "foo": "2"}
 
         # `extra_query` takes priority over `query` when keys clash
-        request = self.client._build_request(
+        request = client._build_request(
             FinalRequestOptions(
                 method="post",
                 url="/foo",
@@ -1514,7 +1555,7 @@ class TestAsyncSwarmsClient:
         ]
 
     @pytest.mark.respx(base_url=base_url)
-    async def test_basic_union_response(self, respx_mock: MockRouter) -> None:
+    async def test_basic_union_response(self, respx_mock: MockRouter, async_client: AsyncSwarmsClient) -> None:
         class Model1(BaseModel):
             name: str
 
@@ -1525,14 +1566,16 @@ class TestAsyncSwarmsClient:
             return_value=httpx.Response(200, json={"foo": "bar"})
         )
 
-        response = await self.client.get(
+        response = await async_client.get(
             "/foo", cast_to=cast(Any, Union[Model1, Model2])
         )
         assert isinstance(response, Model2)
         assert response.foo == "bar"
 
     @pytest.mark.respx(base_url=base_url)
-    async def test_union_response_different_types(self, respx_mock: MockRouter) -> None:
+    async def test_union_response_different_types(
+        self, respx_mock: MockRouter, async_client: AsyncSwarmsClient
+    ) -> None:
         """Union of objects with the same field name using a different type"""
 
         class Model1(BaseModel):
@@ -1545,7 +1588,7 @@ class TestAsyncSwarmsClient:
             return_value=httpx.Response(200, json={"foo": "bar"})
         )
 
-        response = await self.client.get(
+        response = await async_client.get(
             "/foo", cast_to=cast(Any, Union[Model1, Model2])
         )
         assert isinstance(response, Model2)
@@ -1553,7 +1596,7 @@ class TestAsyncSwarmsClient:
 
         respx_mock.get("/foo").mock(return_value=httpx.Response(200, json={"foo": 1}))
 
-        response = await self.client.get(
+        response = await async_client.get(
             "/foo", cast_to=cast(Any, Union[Model1, Model2])
         )
         assert isinstance(response, Model1)
@@ -1561,7 +1604,7 @@ class TestAsyncSwarmsClient:
 
     @pytest.mark.respx(base_url=base_url)
     async def test_non_application_json_content_type_for_json_data(
-        self, respx_mock: MockRouter
+        self, respx_mock: MockRouter, async_client: AsyncSwarmsClient
     ) -> None:
         """
         Response that sets Content-Type to something other than application/json but returns json data
@@ -1578,11 +1621,11 @@ class TestAsyncSwarmsClient:
             )
         )
 
-        response = await self.client.get("/foo", cast_to=Model)
+        response = await async_client.get("/foo", cast_to=Model)
         assert isinstance(response, Model)
         assert response.foo == 2
 
-    def test_base_url_setter(self) -> None:
+    async def test_base_url_setter(self) -> None:
         client = AsyncSwarmsClient(
             base_url="https://example.com/from_init",
             api_key=api_key,
@@ -1594,39 +1637,26 @@ class TestAsyncSwarmsClient:
 
         assert client.base_url == "https://example.com/from_setter/"
 
-    def test_base_url_env(self) -> None:
+        await client.close()
+
+    async def test_base_url_env(self) -> None:
         with update_env(SWARMS_CLIENT_BASE_URL="http://localhost:5000/from/env"):
             client = AsyncSwarmsClient(
                 api_key=api_key, _strict_response_validation=True
             )
             assert client.base_url == "http://localhost:5000/from/env/"
 
-    @pytest.mark.parametrize(
-        "client",
-        [
-            AsyncSwarmsClient(
-                base_url="http://localhost:5000/custom/path/",
-                api_key=api_key,
-                _strict_response_validation=True,
-            ),
-            AsyncSwarmsClient(
-                base_url="http://localhost:5000/custom/path/",
-                api_key=api_key,
-                _strict_response_validation=True,
-                http_client=httpx.AsyncClient(),
-            ),
-        ],
-        ids=["standard", "custom http client"],
-    )
-    def test_base_url_trailing_slash(self, client: AsyncSwarmsClient) -> None:
-        request = client._build_request(
-            FinalRequestOptions(
-                method="post",
-                url="/foo",
-                json_data={"foo": "bar"},
-            ),
-        )
-        assert request.url == "http://localhost:5000/custom/path/foo"
+        # explicit environment arg requires explicitness
+        with update_env(SWARMS_CLIENT_BASE_URL="http://localhost:5000/from/env"):
+            with pytest.raises(ValueError, match=r"you must pass base_url=None"):
+                AsyncSwarmsClient(api_key=api_key, _strict_response_validation=True, environment="production")
+
+            client = AsyncSwarmsClient(
+                base_url=None, api_key=api_key, _strict_response_validation=True, environment="production"
+            )
+            assert str(client.base_url).startswith("https://api.swarms.world")
+
+            await client.close()
 
     @pytest.mark.parametrize(
         "client",
@@ -1645,7 +1675,7 @@ class TestAsyncSwarmsClient:
         ],
         ids=["standard", "custom http client"],
     )
-    def test_base_url_no_trailing_slash(self, client: AsyncSwarmsClient) -> None:
+    async def test_base_url_trailing_slash(self, client: AsyncSwarmsClient) -> None:
         request = client._build_request(
             FinalRequestOptions(
                 method="post",
@@ -1654,6 +1684,7 @@ class TestAsyncSwarmsClient:
             ),
         )
         assert request.url == "http://localhost:5000/custom/path/foo"
+        await client.close()
 
     @pytest.mark.parametrize(
         "client",
@@ -1672,7 +1703,35 @@ class TestAsyncSwarmsClient:
         ],
         ids=["standard", "custom http client"],
     )
-    def test_absolute_request_url(self, client: AsyncSwarmsClient) -> None:
+    async def test_base_url_no_trailing_slash(self, client: AsyncSwarmsClient) -> None:
+        request = client._build_request(
+            FinalRequestOptions(
+                method="post",
+                url="/foo",
+                json_data={"foo": "bar"},
+            ),
+        )
+        assert request.url == "http://localhost:5000/custom/path/foo"
+        await client.close()
+
+    @pytest.mark.parametrize(
+        "client",
+        [
+            AsyncSwarmsClient(
+                base_url="http://localhost:5000/custom/path/",
+                api_key=api_key,
+                _strict_response_validation=True,
+            ),
+            AsyncSwarmsClient(
+                base_url="http://localhost:5000/custom/path/",
+                api_key=api_key,
+                _strict_response_validation=True,
+                http_client=httpx.AsyncClient(),
+            ),
+        ],
+        ids=["standard", "custom http client"],
+    )
+    async def test_absolute_request_url(self, client: AsyncSwarmsClient) -> None:
         request = client._build_request(
             FinalRequestOptions(
                 method="post",
@@ -1681,35 +1740,35 @@ class TestAsyncSwarmsClient:
             ),
         )
         assert request.url == "https://myapi.com/foo"
+        await client.close()
 
     async def test_copied_client_does_not_close_http(self) -> None:
-        client = AsyncSwarmsClient(
+        test_client = AsyncSwarmsClient(
             base_url=base_url, api_key=api_key, _strict_response_validation=True
         )
-        assert not client.is_closed()
+        assert not test_client.is_closed()
 
-        copied = client.copy()
-        assert copied is not client
+        copied = test_client.copy()
+        assert copied is not test_client
 
         del copied
 
         await asyncio.sleep(0.2)
-        assert not client.is_closed()
+        assert not test_client.is_closed()
 
     async def test_client_context_manager(self) -> None:
-        client = AsyncSwarmsClient(
+        test_client = AsyncSwarmsClient(
             base_url=base_url, api_key=api_key, _strict_response_validation=True
         )
-        async with client as c2:
-            assert c2 is client
+        async with test_client as c2:
+            assert c2 is test_client
             assert not c2.is_closed()
-            assert not client.is_closed()
-        assert client.is_closed()
+            assert not test_client.is_closed()
+        assert test_client.is_closed()
 
     @pytest.mark.respx(base_url=base_url)
-    @pytest.mark.asyncio
     async def test_client_response_validation_error(
-        self, respx_mock: MockRouter
+        self, respx_mock: MockRouter, async_client: AsyncSwarmsClient
     ) -> None:
         class Model(BaseModel):
             foo: str
@@ -1719,7 +1778,7 @@ class TestAsyncSwarmsClient:
         )
 
         with pytest.raises(APIResponseValidationError) as exc:
-            await self.client.get("/foo", cast_to=Model)
+            await async_client.get("/foo", cast_to=Model)
 
         assert isinstance(exc.value.__cause__, ValidationError)
 
@@ -1733,7 +1792,6 @@ class TestAsyncSwarmsClient:
             )
 
     @pytest.mark.respx(base_url=base_url)
-    @pytest.mark.asyncio
     async def test_received_text_for_expected_json(
         self, respx_mock: MockRouter
     ) -> None:
@@ -1751,12 +1809,15 @@ class TestAsyncSwarmsClient:
         with pytest.raises(APIResponseValidationError):
             await strict_client.get("/foo", cast_to=Model)
 
-        client = AsyncSwarmsClient(
+        non_strict_client = AsyncSwarmsClient(
             base_url=base_url, api_key=api_key, _strict_response_validation=False
         )
 
-        response = await client.get("/foo", cast_to=Model)
+        response = await non_strict_client.get("/foo", cast_to=Model)
         assert isinstance(response, str)  # type: ignore[unreachable]
+
+        await strict_client.close()
+        await non_strict_client.close()
 
     @pytest.mark.parametrize(
         "remaining_retries,retry_after,timeout",
@@ -1780,17 +1841,14 @@ class TestAsyncSwarmsClient:
         ],
     )
     @mock.patch("time.time", mock.MagicMock(return_value=1696004797))
-    @pytest.mark.asyncio
     async def test_parse_retry_after_header(
-        self, remaining_retries: int, retry_after: str, timeout: float
+        self, remaining_retries: int, retry_after: str, timeout: float, async_client: AsyncSwarmsClient
     ) -> None:
-        client = AsyncSwarmsClient(
-            base_url=base_url, api_key=api_key, _strict_response_validation=True
-        )
+        
 
         headers = httpx.Headers({"retry-after": retry_after})
         options = FinalRequestOptions(method="get", url="/foo", max_retries=3)
-        calculated = client._calculate_retry_timeout(
+        calculated = async_client._calculate_retry_timeout(
             remaining_retries, options, headers
         )
         assert calculated == pytest.approx(
@@ -1812,7 +1870,7 @@ class TestAsyncSwarmsClient:
         with pytest.raises(APITimeoutError):
             await async_client.with_streaming_response.get_root().__aenter__()
 
-        assert _get_open_connections(self.client) == 0
+        assert _get_open_connections(async_client) == 0
 
     @mock.patch(
         "swarms_client._base_client.BaseClient._calculate_retry_timeout",
@@ -1826,7 +1884,7 @@ class TestAsyncSwarmsClient:
 
         with pytest.raises(APIStatusError):
             await async_client.with_streaming_response.get_root().__aenter__()
-        assert _get_open_connections(self.client) == 0
+        assert _get_open_connections(async_client) == 0
 
     @pytest.mark.parametrize("failures_before_success", [0, 2, 4])
     @mock.patch(
@@ -1834,7 +1892,6 @@ class TestAsyncSwarmsClient:
         _low_retry_timeout,
     )
     @pytest.mark.respx(base_url=base_url)
-    @pytest.mark.asyncio
     @pytest.mark.parametrize("failure_mode", ["status", "exception"])
     async def test_retries_taken(
         self,
@@ -1872,7 +1929,6 @@ class TestAsyncSwarmsClient:
         _low_retry_timeout,
     )
     @pytest.mark.respx(base_url=base_url)
-    @pytest.mark.asyncio
     async def test_omit_retry_count_header(
         self,
         async_client: AsyncSwarmsClient,
@@ -1906,7 +1962,6 @@ class TestAsyncSwarmsClient:
         _low_retry_timeout,
     )
     @pytest.mark.respx(base_url=base_url)
-    @pytest.mark.asyncio
     async def test_overwrite_retry_count_header(
         self,
         async_client: AsyncSwarmsClient,
@@ -1961,7 +2016,7 @@ class TestAsyncSwarmsClient:
         )
 
     @pytest.mark.respx(base_url=base_url)
-    async def test_follow_redirects(self, respx_mock: MockRouter) -> None:
+    async def test_follow_redirects(self, respx_mock: MockRouter, async_client: AsyncSwarmsClient) -> None:
         # Test that the default follow_redirects=True allows following redirects
         respx_mock.post("/redirect").mock(
             return_value=httpx.Response(
@@ -1972,14 +2027,14 @@ class TestAsyncSwarmsClient:
             return_value=httpx.Response(200, json={"status": "ok"})
         )
 
-        response = await self.client.post(
+        response = await async_client.post(
             "/redirect", body={"key": "value"}, cast_to=httpx.Response
         )
         assert response.status_code == 200
         assert response.json() == {"status": "ok"}
 
     @pytest.mark.respx(base_url=base_url)
-    async def test_follow_redirects_disabled(self, respx_mock: MockRouter) -> None:
+    async def test_follow_redirects_disabled(self, respx_mock: MockRouter, async_client: AsyncSwarmsClient) -> None:
         # Test that follow_redirects=False prevents following redirects
         respx_mock.post("/redirect").mock(
             return_value=httpx.Response(
@@ -1988,7 +2043,7 @@ class TestAsyncSwarmsClient:
         )
 
         with pytest.raises(APIStatusError) as exc_info:
-            await self.client.post(
+            await async_client.post(
                 "/redirect",
                 body={"key": "value"},
                 options={"follow_redirects": False},
